Add combo chart type, site overhaul, and v0.0.0.5 release
Combo chart (7th certified type):
- Overlay column + line marks on shared Cartesian axes with dual y-axis support
- Python and Go renderers with byte-identical parity
- 4 golden SVGs (basic, dark, dual-axis, adversarial)
- Schema, validation, browser qualification, and fuzz coverage

Shared fixes discovered during combo work:
- Fix ypix2 mapping secondary-axis values to wrong dimension (cartesian)
- Fix _nice_num returning int instead of float (util.py)
- Add monotone spline curve support to area charts (area.go)
- Harden verify/cli.py with type annotations and context managers

Site overhaul:
- Redesign landing page with stats bar, problem/solution cards, JSON-LD
- Add SEO meta (canonical, OG image, Twitter cards, sitemap, CSP headers)
- Add accessibility features (skip-link, focus-visible, aria-labels)
- Enhance demo pages (benchmarks bar-meters, verify explanations)
- Add 404 page and og-preview.svg

Docs and CI:
- Update capability matrix, one-pager, backlog, and constitution for combo
- Tighten content-draft articles with governed frontmatter
- Add combo to CI wheel smoke test and cross-language regression loop
- Add .coverage to .gitignore

+++ libs/python/stonecharts/util.py
@@ -18,7 +18,7 @@ def _nice_num(x: float, round_: bool) -> float:
         nf = 1 if frac < 1.5 else 2 if frac < 3 else 5 if frac < 7 else 10
     else:
         nf = 1 if frac <= 1 else 2 if frac <= 2 else 5 if frac <= 5 else 10
-    return nf * (10**exp)
+    return float(nf * (10**exp))
 
 
 def nice_ticks(lo: float, hi: float, target: int = 6) -> tuple[float, float, list[float]]:
